+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @lmarcel/book
 
+## 0.2.0
+
+### Minor Changes
+
+- Changing primary theme, adding tokens and creating custom theme class
+
+### Patch Changes
+
+- Updated dependencies
+  - @lmarcel/highlight@1.5.0
+
 ## 0.1.1
 
 ### Patch Changes